airflow/requirements.txt
python-dotenv
tiktoken
boto3
docling
openai
apache-airflow==2.10.4
apache-airflow-providers-amazon
apache-airflow-providers-smtp
mistralai
pinecone
spacy